requirements.txt
ccxt==4.1.13
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
solana==0.30.2
solders==0.18.1
base58==2.1.1
flask==3.1.0
flask-socketio==5.5.1
requests-cache==1.1.1
pyth-client==0.8.0
pythclient==0.1.12
aiohttp==3.9.3
python-binance==1.0.19
coingecko-api==1.0.4
pycoingecko==3.1.0
gunicorn==21.2.0
eventlet==0.35.2 